+++ devo/sender/scripts/sender_cli.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """CLI for use Devo Sender from shell command line."""