+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [No version upgraded] - 2020-04-29
+### Added
+ * Added examples
+ 
+### Fixed
+ * Fixed shebangs
+ * Fixed typo in README
+ * Remove one blank in Sender
+
 ## [3.3.0] - 2020-04-14
 ### Added
  * Support for Python 3.8